backend/requirements.txt
fastapi
uvicorn
motor
pydantic
python-dotenv
apscheduler
exponent-server-sdk
httpx
